Fix EC driver bugs, suspend bugs.
ACPI: Fix various bugs with the EC driver.
x86_64: Restore NVS memory, I/O APIC state, and HPET state on wake
uACPI: Remove un needed kernel api
ACPI: Before reboot/shutdown, make it suspend the scheduler, as opposed to halting all CPUs.
Misc: Add experimental OBOS_ENABLE_LOCK_PROFILING option to measure how long mutexes and spinlocks are taken
Misc: Fix build system regressions.

+++ src/uACPI/uacpi_arch_helpers.h
@@ -20,7 +20,7 @@
 #endif
 
 typedef uint8_t uacpi_cpu_flags;
-typedef uint32_t uacpi_thread_id;
+typedef uintptr_t uacpi_thread_id;
 
 #ifndef UACPI_THREAD_ID_NONE
 #    define UACPI_THREAD_ID_NONE ((uacpi_thread_id)-1)
